Restructure baseline, there are still some loose bits, but I'm now happy with the outline.
